Handle case of root env poisoning :-)
Become or not become, that is the question.

Whenever a task with become: yes is executed first, the gathered facts (i.e.,
ansible_env.HOME) will be root's instead of the main user.

This plagues several of the user-oriented roles, and this is a workaround for
it.

OTOH, we can simply set gather_facts: no by default, and add gather_facts: yes
to whatever plays need an updated env.